+++ routes/user/login.js
@@ -4,7 +4,7 @@ async function login(req, res) {
     var username = req.body.username,
         password = req.body.password;
 
-    var public_path = __dirname.replace("routes", "");
+    var public_path = __dirname.replace("routes", "").replace("user", "");
     var main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
         downloads_path = main_path + username + "_Downloads";
